Move receive page into tab
